+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## [9.0.0] - 2024-11-13
+
 ### Added
 
 - `nylasGrantId` and `nylasConfigurationId` to `FormSubmissionEventConfiguration`